% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCFIC.R
\name{eccfic}
\alias{eccfic}
\title{ECCFIC Sample Estimates}
\usage{
eccfic(
  x,
  y,
  est = "kernel",
  ns = NULL,
  kernel = "gaussian",
  bw = "default",
  sigma = "default",
  index = 1,
  wt = F
)
}
\arguments{
\item{x}{Data of first sample serving as the response}

\item{y}{Data of second sample conditioned on}

\item{est}{The estimation method for the conditional mean. Set
to 'slice' for slicing estimation or to 'kernel' for kernel.
regression estimation. See 'Details' for additional information}

\item{ns}{Number of slices. Defaults to NULL and must be specified when
method is set to 'slice' for a continuous y}

\item{kernel}{A kernel embedding to use for x, 'gaussian' or 'distance}

\item{bw}{Bandwidth of the gaussian smoothing kernel applied on y, bandwidths
suggested by Silverman (1986) are used unless otherwise specified.}

\item{sigma}{Bandwidth, if a 'gaussian' kernel is used for x, default is
heuristic median pairwise distances of x.}

\item{index}{Exponent on distance in (0,2], if a 'distance' kernel is used
for x. Defaults to 1}

\item{wt}{Use weight function or not, if TRUE, f^2 is used by default}
}
\value{
\code{eccfic} returns a list with components \describe{
\item{eccfic}{sample ECCFIC estimate of \eqn{H_k^2(X|Y)}}
\item{eccficVar}{sample ECCFIC variance estimate}
\item{eccficCor}{sample correlation type statistic } }
}
\description{
Computes Expected Conditional Characteristic Function-based
Independence Criterion (ECCFIC) statistics, which are new multivariate measures of
independence.
}
\details{
When estimation method (est) is set to 'slice', the ECCFIC measure is computed by slicing
on Y in which case Y can originally be categorical or continuous. Estimating ECCFIC
is straightforward if Y is categorical.  For a continuous Y, we change Y into finite
categories with 'ns' levels. Slicing in multivariate and high-dimensional situations
is very challenging. So, for a continuous multivariate Y, we recommend using the kernel approach
or slicing Y before hand to be able to use the slicing approach. For help on slicing, please see
the works by \insertCite{zhu2010sufficient}{eccficm},
\insertCite{li2008projective}{eccficm}, and
\insertCite{cook2014fused}{eccficm}.

Kernel regression is an alternative approach to use for a continuous Y. To use
the kernel regression method simply set \code{est} to 'kernel'.

Refer to section 4 of \insertCite{yin2019new}{eccficm} for more information on
the estimation approaches.

Though ECCFIC is not very sensitive to the number of slices, it is suggested that
each slice should have at least 5 and at most n/2 data points.

ECD \insertCite{yin2019new}{eccficm}, \code{\link{ecd}}, belongs to the family of generalized
ECCFIC measures. In fact, \code{ecdCov^2} = 2*\code{eccfic}, when both measures
are estimated by the slicing method and a distance kernel is used for the ECCFIC.
}
\note{
See \code{\link{eccfic.test}} for a test of multivariate independence based on the
ECCFIC statistic as well as the correlation type statistic.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100)

# Using kernel regression estimation method
eccfic(x,y, est="kernel")

# Using slicing estimation method for a continuous Y
eccfic(x,y, est="slice", ns=5)

# Using slicing method for a categorical Y
y <- rep(1:4, 25)
x <- y + rnorm(100)
# don't need to provide 'ns' because Y is already categorical
eccfic(x,as.factor(y),est='slice',kernel='distance')

}
\references{
\insertRef{ke2019expected}{eccficm}
\url{https://www.tandfonline.com/doi/abs/10.1080/01621459.2019.1604364}

\insertRef{zhu2010sufficient}{eccficm}

\insertRef{li2008projective}{eccficm}

\insertRef{cook2014fused}{eccficm}
}
\seealso{
\code{\link{ecd}} \code{\link{ecd.test}}
}
\author{
Chenlu Ke (\email{chenlu.ke@uky.edu}) and Xiangrong Yin
}
