% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCFIC_test.R
\name{eccfic.test}
\alias{eccfic.test}
\title{ECCFIC Independence Test}
\usage{
eccfic.test(
  x,
  y,
  est = "kernel",
  ns = NULL,
  kernel = "gaussian",
  B = "default",
  ts = "eccficCor",
  bw = "default",
  sigma = "default",
  index = 1,
  wt = F
)
}
\arguments{
\item{x}{Data of first sample serving as the response}

\item{y}{Data of second sample conditioned on}

\item{est}{The estimation method for the conditional mean. Set
to 'slice' for slicing estimation or to 'kernel' for kernel.
regression estimation. See 'Details' for additional information}

\item{ns}{Number of slices. Defaults to NULL and must be specified when
method is set to 'slice' for a continuous y}

\item{kernel}{A kernel embedding to use for x, 'gaussian' or 'distance}

\item{B}{number of permutation replicates}

\item{ts}{the test statistic, 'eccfic' or 'eccficCor'}

\item{bw}{Bandwidth of the gaussian smoothing kernel applied on y, bandwidths
suggested by Silverman (1986) are used unless otherwise specified.}

\item{sigma}{Bandwidth, if a 'gaussian' kernel is used for x, default is
heuristic median pairwise distances of x.}

\item{index}{Exponent on distance in (0,2], if a 'distance' kernel is used
for x. Defaults to 1}

\item{wt}{Use weight function or not, if TRUE, f^2 is used by default}
}
\value{
\code{eccfic.test} returns a list containing the following components:
\describe{
\item{estimation.method}{a character string indicating the method used to compute the ECCFIC statistic}
\item{method}{a character string describing the type of ECCFIC test performed}
\item{statistic}{the value of the ECCFIC test statistic}
\item{p.value}{approximate p-value of the test based on a permutation test}
\item{alternative }{a character string describing the alternative hypothesis}
\item{estimates}{vector containing estimates of H2(X|Y), H2(X|X), and Rho(X|Y)}
\item{data.name}{a character string containing the names of the data used}
}
}
\description{
Performs a permutation test of independence between
two random vectors based on the ECCFIC statistic or the correlation type statistic.
}
\details{
For the permutation test, if no value is given for 'B', (200+500/n) replicates
are used by default, where n is the sample size. At least 99 and at most 999 random permutations
should be sufficient practically.

When estimation method (est) is set to 'slice', the ECCFIC measure is computed by slicing
on Y in which case Y can originally be categorical or continuous. Estimating ECCFIC
is straightforward if Y is categorical.  For a continuous Y, we change Y into finite
categories with 'ns' levels. Slicing in multivariate and high-dimensional situations
is very challenging. So, for a continuous multivariate Y, we recommend using the kernel approach
or slicing Y before hand to be able to use the slicing approach. For help on slicing, please see
the works by \insertCite{zhu2010sufficient}{eccficm},
\insertCite{li2008projective}{eccficm}, and
\insertCite{cook2014fused}{eccficm}.

Kernel regression is an alternative approach to use for a continuous Y. To use
the kernel regression method simply set \code{est} to 'kernel'.

Refer to section 4 of \insertCite{yin2019new}{eccficm} for more information on
the estimation approaches.

Though ECCFIC is not very sensitive to the number of slices, it is suggested that
each slice should have at least 5 and at most n/2 data points.

ECD \insertCite{yin2019new}{eccficm}, \code{\link{ecd}}, belongs to the family of generalized
ECCFIC measures. In fact, \code{ecdCov^2} = 2*\code{eccfic}, when both measures
are estimated by the slicing method and a distance kernel is used for the ECCFIC.
}
\examples{
x <- rnorm(100)
y <- x + rnorm(100)

# Using Gaussian Kernel Regression estimation method
eccfic.test(x,y) # based on default permutations
eccfic.test(x,y,B=199) # based on 199 permutations

# Using slicing estimation method on a continuous Y
eccfic.test(x,y,est='slice',ns=5,kernel='distance') # based on default permutations
eccfic.test(x,y,est='slice',ns=5,kernel='distance',B=199) # based on 199 permutations

# Using slicing method for a categorical Y
y <- rep(1:4, 25)
x <- y + rnorm(100)
# don't need to provide 'ns' because Y is already categorical
eccfic.test(x,as.factor(y),est='slice',kernel='distance')

}
\references{
\insertRef{ke2019expected}{eccficm}
\url{https://www.tandfonline.com/doi/abs/10.1080/01621459.2019.1604364}
}
\seealso{
\code{\link{eccfic}} \code{\link{ecd.test}}
}
\author{
Chenlu Ke (\email{chenlu.ke@uky.edu}) and Xiangrong Yin
}
