% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECD.R
\name{ecd}
\alias{ecd}
\title{ECD Sample Estimates}
\usage{
ecd(x, y, est = "slice", ns = 5, bw = "default", index = 1)
}
\arguments{
\item{x}{Data of first sample serving as the response}

\item{y}{Data of second sample conditioned on}

\item{est}{The estimation method (options: 'slice', 'kernel.epa', 'kernel.gau').
See 'Details' for additional information}

\item{ns}{Number of slices. Defaults to NULL and must be specified when method
is set to 'slice' for a continuous y}

\item{bw}{Bandwidth of the kernel function, default bandwidths  are used
unless otherwise specified}

\item{index}{Exponent on Euclidean distance in (0,2], with default value of 1}
}
\value{
\code{ecd} returns a list with components
\describe{
\item{ecdCov}{sample ECD covariance type statistic denoted \eqn{C_n(X|Y)}}
\item{ecdCor}{sample ECD correlation type statistic}
\item{ecdVarX}{sample ECD variance type statistic of X}
}
}
\description{
Computes Expectation of the Conditional Difference (ECD) statistics.
These covariance and correlation type statistics are multivariate measures of
independence.
}
\details{
When estimation method (est) is set to 'slice', the ECD measure is computed by slicing
on Y in which case Y can originally be categorical or continuous. Estimating ECD
is straightforward if Y is categorical. For a continuous Y, we first change Y into
finite categories with 'ns' levels. Slicing in multivariate and high-dimensional situations
is very challenging. So, for a continuous multivariate Y, we recommend using the kernel approach
or slicing Y before hand to be able to use the slicing approach. For help on slicing,
please see the works by
\insertCite{zhu2010sufficient}{eccficm}, \insertCite{li2008projective}{eccficm},
and \insertCite{cook2014fused}{eccficm}.

Kernel estimation is an alternative approach to use for a continuous Y. The
available method options, 'kernel.epa' and 'kernel.gau', make use of the
Epanechnikov kernel function and Gaussian kernel function, respectively.

Refer to section 4 of \insertCite{yin2019new}{eccficm} for more information on
the estimation approaches.

ECD \insertCite{yin2019new}{eccficm}, \code{\link{ecd}}, belongs to the family of generalized
ECCFIC measures. In fact, \code{ecdCov^2} = 2*\code{eccfic}, when both measures
are estimated by the slicing method and a distance kernel is used for the ECCFIC.
}
\note{
See \code{\link{ecd.test}} for a test of multivariate independence based on the
covariance and correlation type statistics.
}
\examples{
library(MASS)
n <- 30; p <- 5; q <- 1
mu <- rep(0,(p+q))
sigma <- diag((p+q))
DATA <- mvrnorm(n, mu, sigma)
X <- DATA[,1:p]
Y <- DATA[,(p+1):(p+q)]

# using slicing estimation method on a continuous Y
ecd(x = X, y = Y, est = "slice", ns = 5)

# using Gaussian kernel estimation method
ecd(x = X, y = Y, est = "kernel.gau")

# using Epanechnikov kernel estimation method
ecd(x = X, y = Y, est = "kernel.epa")

#This example shows how to directly compute the ECD measures if Y is a
#categorical variable
library(boot)
#gravity data is in the boot package,there are 81 measurements in a series of
#eight experiments
#gravity$series shows the categories 1-8 of the observations.
str(gravity)
ecd(gravity$g, gravity$series, est="slice")

}
\references{
\insertRef{yin2019new}{eccficm}
\url{http://www3.stat.sinica.edu.tw/ss_newpaper/SS-2017-0538_na.pdf}

\insertRef{zhu2010sufficient}{eccficm}

\insertRef{li2008projective}{eccficm}

\insertRef{cook2014fused}{eccficm}
}
\seealso{
\code{\link{eccfic}}  \code{\link{eccfic.test}}
}
\author{
Qinqcong Yuan (\email{yuanq3@miamioh.edu}) and Xiangrong Yin
}
