% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECD_test.R
\name{ecd.test}
\alias{ecd.test}
\title{ECD Independence Test}
\usage{
ecd.test(
  x,
  y,
  est = "slice",
  ns = 5,
  ts = "ecdcor",
  B = "default",
  bw = "default",
  index = 1
)
}
\arguments{
\item{x}{Data of first sample serving as the response}

\item{y}{Data of second sample conditioned on}

\item{est}{The estimation method (options: 'slice', 'kernel.epa', 'kernel.gau').
See 'Details' for additional information}

\item{ns}{Number of slices. Defaults to NULL and must be specified when method
is set to 'slice' for a continuous y}

\item{ts}{the test statistic, 'ecdcor' or 'ecdcov'}

\item{B}{number of permutation replicates}

\item{bw}{Bandwidth of the kernel function, default bandwidths  are used
unless otherwise specified}

\item{index}{Exponent on Euclidean distance in (0,2], with default value of 1}
}
\value{
\code{ecd.test} returns a list containing the following components:
\describe{
\item{estimation.method}{a character string indicating the method used to compute the ECD measure}
\item{method}{a character string describing the type of ECD test performed}
\item{statistic}{the value of the ecd test statistic}
\item{p.value}{approximate p-value of the test based on a permutation test}
\item{alternative }{a character string describing the alternative hypothesis}
\item{estimates}{vector containing ECD statistics, ecdCov, ecdCor, and ecdVarX}
\item{data.name}{a character string containing the names of the data used}
}
}
\description{
Performs a permutation test of independence between
two random vectors based on the ECD covariance and correlation type statistics.
}
\details{
For the permutation test, if no value is given for 'B', (200+500/n) replicates
are used by default, where n is the sample size. At least 99 and at most 999 random permutations
should be sufficient practically.

When estimation method (est) is set to 'slice', the ECD measure is computed by slicing
on Y in which case Y can originally be categorical or continuous. Estimating ECD
is straightforward if Y is categorical. For a continuous Y, we first change Y into
finite categories with 'ns' levels. Slicing in multivariate and high-dimensional situations
is very challenging. So, for a continuous multivariate Y, we recommend using the kernel approach
or slicing Y before hand to be able to use the slicing approach. For help on slicing,
please see the works by
\insertCite{zhu2010sufficient}{eccficm}, \insertCite{li2008projective}{eccficm},
and \insertCite{cook2014fused}{eccficm}.

Kernel estimation is an alternative approach to use for a continuous Y. The
available method options, 'kernel.epa' and 'kernel.gau', make use of the
Epanechnikov kernel function and Gaussian kernel function, respectively.

Refer to section 4 of \insertCite{yin2019new}{eccficm} for more information on
the estimation approaches.
}
\examples{
library(MASS)
n <- 30; p <- 5; q <- 1
mu <- rep(0,(p+q))
sigma <- diag((p+q))
DATA <- mvrnorm(n, mu, sigma)
X <- DATA[,1:p]
Y <- DATA[,(p+1):(p+q)]

# Using slicing estimation method on a continuous Y
ecd.test(X, Y, est = "slice", ns = 5,ts="ecdcor", B = "default")

# Using Gaussian kernel estimation method
ecd.test(X, Y, est = "kernel.gau", ts="ecdcor", B = "default")

# Using Epanechnikov kernel estimation method
ecd.test(X, Y, est = "kernel.epa", ts="ecdcor", B = "default")

# This example shows how to directly perform the test if Y is a categorical variable
library(boot)
#gravity data is in the boot package,there are 81 measurements in a series
# of eight experiments
#gravity$series shows the categories 1-8 of the observations. 999 permutations
# are used to compute p-value.
str(gravity)
ecd.test(gravity$g, gravity$series, est="slice", ts="ecdcor", B=999)

}
\references{
\insertRef{yin2019new}{eccficm}
}
\seealso{
\code{\link{ecd}} \code{\link{eccfic.test}}
}
\author{
Qinqcong Yuan (\email{yuanq3@miamioh.edu}) and Xiangrong Yin
}
