% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdcor_test.R
\name{ecdcor.test}
\alias{ecdcor.test}
\title{ECD Independence Test}
\usage{
ecdcor.test(x, y, method, ns = NULL, B = "default", bw = "default", index = 1)
}
\arguments{
\item{x}{data of first sample serving as the response}

\item{y}{data of second sample conditioned on}

\item{method}{ECD estimation method (options: slice, kernel.epa, kernel.gau)}

\item{ns}{number of slices. Defaults to NULL and must be specified when method is set to 'slice' for a continuous y}

\item{B}{number of permutation replicates}

\item{bw}{bandwidth of the kernel function, default
bandwidths  are used unless otherwise specified}

\item{index}{exponent on Euclidean distance, in (0,2]. Defaults to 1}
}
\value{
\code{ecdcor.test} returns a list containing the following components:
\describe{
\item{estimation.method}{a character string indicating the method used to compute the ECD measure}
\item{method}{a character string describing the type of ECD test performed}
\item{statistic}{the value of the ecd test statistic}
\item{p.value}{approximate p-value of the test based on a permutation test}
\item{alternative }{a character string describing the alternative hypothesis}
\item{estimates}{vector containing ecdCov, ecdCor, and ecdVarX}
\item{data.name}{a character string containing the names of the data used}
}
}
\description{
Performs a permutation test of independence based on
the distance correlation type of measure
}
\examples{
library(MASS)
n=30
p=5
q=1
mu=rep(0,(p+q))
sigma=diag((p+q))
DATA=mvrnorm(n, mu, sigma)
X=DATA[,1:p]
Y=DATA[,(p+1):(p+q)]
# using slicing estimation method on a continuous Y
ecdcor.test(x = X, y = Y, method = "slice", B = "default", ns = 5)
# using Gaussian kernel estimation method
ecdcor.test(x = X, y = Y, method = "kernel.gau", B = "default")
# using Epanechnikov kernel estimation method
ecdcor.test(x = X, y = Y, method = "kernel.epa", B = "default")

#This example shows how to directly perform the test if Y is a categorical
#variable
library(boot)
#gravity data is in the boot package,there are 81 measurements in a series
# of eight experiments
#gravity$series shows the categories 1-8 of the observations. 999 permutations
# are used to compute p-value.
str(gravity)
ecdcor.test(gravity$g, gravity$series, method="slice", B=999)
}
\seealso{
@seealso \code{\link{ecd}} \code{\link{ecdcov.test}}
}
