% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdmeasures.R
\docType{package}
\name{ecdmeasures}
\alias{ecdmeasures}
\title{ecdmeasures: A package for computing new distance correlation/covariance
measures for testing independence between two random vectors.}
\description{
The ecdmeasures package provides three main functions:
ecd, ecdcor and ecdcov.test
}
\section{ecdmeasures functions}{

The ecd function computes the distance covariance and distance correlation type of measures

The ecdcor.test function performs a permutation test of independence based on
the distance correlation type of measure

The ecdcov.test function performs a permutation test of independence based on
the distance covariance type of measure
}

