% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscreen.R
\name{fscreen}
\alias{fscreen}
\title{Sure Independence Screening}
\usage{
fscreen(X, Y, method = "ECD-SIS", est = "default", ns = NULL)
}
\arguments{
\item{X}{A data matrix or object of potential predictors. Can be of any dimension.}

\item{Y}{A vector of response. This implementation only supports a univariate
response.}

\item{method}{The screening procedure based on a particular correlation
measure. Four methods are available: "SIS", "DC-SIS", "ECD-SIS", and "ECCFIC-SIS".
The method names are case insensitive meaning one can either type DC-SIS or dc-sis
for the same method. See the 'Details' section for more information.}

\item{est}{The estimation procedure for ECD or ECCFIC. Use 'slice' for the
slicing approach and 'kernel' for the kernel estimation procedure.}

\item{ns}{The number of slices to use when the slicing estimation procedure
is specified.}
}
\value{
\code{fscreen} returns a list object containing the following:
\describe{
\item{utility}{ A vector of the marginal correlations between each predictor/feature
and the response Y. Each value is a measure of the estimated strength of
association with Y.}
\item{rank}{A vector of the position for each predictor/feature in terms of their
predictive power. Ranks are assigned from 1 to the number of columns in the feature
matrix X, with 1 indicating the best predictive performance.
The lower the rank the more active the predictor at that index is in explaining/predicting Y.}
}
}
\description{
Performs the sure independence screening using correlation learning
between each predictor and the response.
}
\section{Screening methods}{

\itemize{
\item{SIS: Sure Independence Screening based on the Pearson correlation coefficient}
\item{DC-SIS: Distance Correlation Sure Independence Screening. Uses dcor() from
the energy package.}
\item{ECD-SIS: Expectation of the Conditional Difference Sure Independence Screening.
When 'est' is set default the slicing method is used, otherwise, the kernel
estimation method can be specified via the 'est' parameter.}
\item{ECCFIC-SIS: The default method uses the kernel regression estimation.
Otherwise, the slicing procedure can be specified through the 'est' parameter.}
}
}

