% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dat.R
\name{generateData}
\alias{generateData}
\title{Data Generation Function}
\usage{
generateData(
  n = 200,
  p = 2000,
  rho = 0,
  model = 1,
  sigma = 1,
  penalize = T,
  beta = c(2.01, 3.98, -2.51, -1.79)
)
}
\arguments{
\item{n}{The number of observations or sample size.}

\item{p}{Dimension of the data matrix of predictors to be generated.}

\item{rho}{Correlation coefficient between predictors.}

\item{model}{A numeric value representing the model for the response.
See the 'Details' section for available options.}

\item{sigma}{The variance of the error term. Defaults to 1.}

\item{penalize}{A boolean indicating whether the penalizing constants should
be used or not.}

\item{beta}{A vector of model beta coefficients. Length must correspond to the
number of beta coefficients in the model. It is only used when the fixed coefficient
models numbered 5 - 8 are desired.
be used or not. These constants are meant to challenge the screening procedures.}
}
\value{
\code{generateData()} returns a list object of the following:
\describe{
\item{X}{A data matrix or object of predictors of dimension p.}
\item{Y}{A vector of response. This implementation only generates a univariate
response.}
}
}
\description{
Generates data for demonstrating the performance of Sure Independence Screening (SIS)
procedures.
}
\details{
The response is simulated based on models provided in example 3.1 of
\insertCite{li2012feature}{eccficm}. The available models are
numbered 1 - 4, from first to last in the order as they appear in the paper.
}
\references{
\insertRef{li2012feature}{eccficm}
}
\author{
William Agyapong and Qingcong Yuan (\email{yuanq3@miamioh.edu})
}
