% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECD.R
\name{ecd}
\alias{ecd}
\title{ECD Sample Estimates}
\usage{
ecd(x, y, method, ns = NULL, bw = "default", index = 1)
}
\arguments{
\item{x}{data of first sample serving as the response}

\item{y}{data of second sample conditioned on}

\item{method}{ECD estimation method (options: slice, kernel.epa, kernel.gau)}

\item{ns}{number of slices. Defaults to NULL and must be specified when method is set to 'slice' for a continuous y}

\item{bw}{bandwidth of the kernel function, default
bandwidths  are used unless otherwise specified}

\item{index}{exponent on Euclidean distance, in (0,2]. Defaults to 1}
}
\value{
\code{ecd} returns a list with components
\describe{
\item{ecdCov}{sample covariance type statistic denoted \eqn{C_n(X|Y)}}
\item{ecdCor}{sample correlation type statistic}
\item{ecdVarX}{ECD variance of x sample}
}
}
\description{
Computes distance covariance and distance correlation type statistics, which are multivariate measures of dependence.
}
\examples{
library(MASS)
n=30
p=5
q=1
mu=rep(0,(p+q))
sigma=diag((p+q))
DATA=mvrnorm(n, mu, sigma)
X=DATA[,1:p]
Y=DATA[,(p+1):(p+q)]
# using slicing estimation method on a continuous Y
ecd(x = X, y = Y, method = "slice", ns = 5)
# using Gaussian kernel estimation method
ecd(x = X, y = Y, method = "kernel.gau")
# using Epanechnikov kernel estimation method
ecd(x = X, y = Y, method = "kernel.epa")

#This example shows how to directly compute the ECD measures if Y is a
#categorical variable
library(boot)
#gravity data is in the boot package,there are 81 measurements in a series of
#eight experiments
#gravity$series shows the categories 1-8 of the observations. 999 permutations
#are used to compute p-value.
str(gravity)
ecd(gravity$g, gravity$series, method="slice")
}
\seealso{
\code{\link{ecdcor.test}} \code{\link{ecdcov.test}}
}
